# Conflicts:
#	.gitignore
#	README-en_US.md
#	README-zh_TW.md
#	README.md

+++ app/src/main/java/com/imcys/bilibilias/base/model/user/DownloadTaskBean.kt
@@ -1,14 +1,19 @@
 package com.imcys.bilibilias.base.model.user
 
-import com.imcys.bilibilias.home.ui.model.*
+import com.imcys.bilibilias.home.ui.model.BangumiPlayBean
+import com.imcys.bilibilias.home.ui.model.BangumiSeasonBean
+import com.imcys.bilibilias.home.ui.model.DashBangumiPlayBean
+import com.imcys.bilibilias.home.ui.model.DashVideoPlayBean
+import com.imcys.bilibilias.home.ui.model.VideoPageListData
+import com.imcys.bilibilias.home.ui.model.VideoPlayBean
 
 data class DownloadTaskDataBean(
-    val cid: Int,
+    val cid: Long,
     val pageTitle: String,
     val bvid: String,
-    //分辨率
+    // 分辨率
     val qn: String,
-    //视频获取方式选择
+    // 视频获取方式选择
     val fnval: String = "80",
     //
     val platform: String = "pc",
@@ -19,4 +24,4 @@ data class DownloadTaskDataBean(
     val videoPageDataData: VideoPageListData.DataBean? = null,
     val bangumiSeasonBean: BangumiSeasonBean.ResultBean.EpisodesBean? = null,
 
-    )
+)
