+++ app/src/main/java/com/imcys/bilibilias/home/ui/adapter/BangumiSubsectionAdapter.kt
@@ -7,13 +7,11 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.imcys.bilibilias.R
 import com.imcys.bilibilias.databinding.ItemBangumiSubsectionBinding
-import com.imcys.bilibilias.databinding.ItemSubsectionBinding
 import com.imcys.bilibilias.home.ui.model.BangumiSeasonBean
-import com.imcys.bilibilias.home.ui.model.VideoPageListData
 
 class BangumiSubsectionAdapter(
     val datas: MutableList<BangumiSeasonBean.ResultBean.EpisodesBean>,
-    val defaultCid: Int,
+    val defaultCid: Long,
     val onClickMethod: (datas: BangumiSeasonBean.ResultBean.EpisodesBean, position: Int) -> Unit,
 ) : RecyclerView.Adapter<BangumiSubsectionAdapter.ViewHolder>() {
